pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-resume-analyzer"
version = "0.1.0"
description = "Agentic RAG Chatbot for resume-vacancy matching"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-ollama>=0.3",
    "langchain-chroma>=0.2",
    "chromadb>=0.6",
    "streamlit>=1.41",
    "sentence-transformers>=3.4",
    "pydantic>=2.10",
    "python-dotenv>=1.0",
    "pypdf>=5.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "locust>=2.32",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
